+++ CHANGELOG.md
@@ -7,11 +7,11 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 ## [3.7.7]
 ### Added
-- Add configurations to enable/disable order status change upon Capture and Refund from AltaPay grid.
+- Sync missing order data from gateway.
 
 ## [3.7.6]
 ### Added
-- Sync missing order data from gateway.
+- Add configurations to enable/disable order status change upon Capture and Refund from AltaPay grid.
 
 ## [3.7.5]
 ### Added