+++ CHANGELOG.md
@@ -8,6 +8,8 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 ## [3.7.7]
 ### Added
 - Sync missing order data from gateway.
+### Fixed
+- Fix: Duplicate ok callback triggers order status change.
 
 ## [3.7.6]
 ### Added